remove all combos except jk = Esc.
After a lot of testing, none of the others were as fast and reliable as
the non-combo alternatives, which also all used just two keys.

Even the Esc combo could probably go, as it mostly covered by defining
jk=Escape in Vim.

README and layout images have not yet been updated to reflect this.